HWKMETRICS-99: allow Hawkular metrics to start and then wait for a Cassandra cluster to become available rather than just failing.
