+++ api/metrics-api-jaxrs/src/main/java/org/hawkular/metrics/api/jaxrs/MetricsServiceProducer.java
@@ -28,11 +28,14 @@ import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.inject.Produces;
 import javax.inject.Inject;
 
+import com.datastax.driver.core.Session;
+import com.google.common.util.concurrent.FutureCallback;
 import org.hawkular.metrics.api.jaxrs.config.Configurable;
 import org.hawkular.metrics.api.jaxrs.config.ConfigurationProperty;
 import org.hawkular.metrics.api.jaxrs.util.Eager;
 import org.hawkular.metrics.core.api.MetricsService;
-import org.hawkular.metrics.core.impl.HawkularMetrics;
+import org.hawkular.metrics.core.impl.cassandra.CassandraSession;
+import org.hawkular.metrics.core.impl.cassandra.MetricsServiceCassandra;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -70,14 +73,28 @@ public class MetricsServiceProducer {
     @Produces
     public MetricsService getMetricsService() {
         if (metricsService == null) {
-            HawkularMetrics.Builder metricsServiceBuilder = new HawkularMetrics.Builder();
+            metricsService = new MetricsServiceCassandra();
             Map<String, String> options = new HashMap<>();
             options.put("cqlport", cqlPort);
             options.put("nodes", nodes);
             options.put("keyspace", keyspace);
 
-            metricsServiceBuilder.withOptions(options);
-            metricsService = metricsServiceBuilder.build();
+            CassandraSession.Builder cassandraSessionBuilder = new CassandraSession.Builder();
+            cassandraSessionBuilder.withOptions(options);
+
+            cassandraSessionBuilder.withInitializationCallback(new FutureCallback<Session>() {
+                @Override
+                public void onSuccess(Session session) {
+                    metricsService.startUp(session);
+                }
+
+                @Override
+                public void onFailure(Throwable t) {
+                    throw new RuntimeException("Error trying to get the Cassandra Session", t);
+                }
+            });
+
+            cassandraSessionBuilder.build();
         }
 
         return metricsService;
